divide GetMetricData related code
